+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed: deleting tag crashes app
 * Fixed tray icon behaviour in some special cases (show trayicon now defaults to true)
 * Added new menu option Help/Getting started
+* Restored option to open notes in new tab
 
 
 ### v2.1.0-beta-1