+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 * [Binary releases](https://github.com/robert7/nixnote2/releases)
 
 ## 2018-08
+### v2.1.0-beta-3
+
 ### v2.1.0-beta-2
 * Fixed european date formats (wrong order of day/month)
 * Updated translation files with string from source code